Add reporting
Also major refactor:

Make TimeTracker a "dumb" class which is injected an object to interact
with the database. Then we create ConsoleTimeTracker which creates a
new connection/transaction for each method invoked from IPython, and
displays query results tabulated.

This way we should also be able to create a TestingTimeTracker which
creates its own transient database for tests and rolls back after each
test.